+++ packages/templates/typescript-multiple/src/config.ts
@@ -1,12 +1,14 @@
 import * as type from './type.handlebars';
-import * as enumTemplate from './enum.handlebars';
+import * as enumTemplate from '../../typescript/src/enum.handlebars';
 import * as scalar from './scalar.handlebars';
 import * as union from './union.handlebars';
 import * as operation from './operation.handlebars';
 import * as fragment from './fragment.handlebars';
-import * as selectionSet from './selection-set.handlebars';
+import * as selectionSet from '../../typescript/src/selection-set.handlebars';
 import * as fragments from './fragments.handlebars';
 import { EInputType, GeneratorConfig } from 'graphql-codegen-core';
+import { getType } from '../../typescript/src/helpers/get-type';
+import { getOptionals } from '../../typescript/src/helpers/get-optionals';
 
 export const config: GeneratorConfig = {
   inputType: EInputType.MULTIPLE_FILES,
@@ -29,6 +31,10 @@ export const config: GeneratorConfig = {
     Float: 'number',
     Boolean: 'boolean',
     ID: 'string'
+  },
+  customHelpers: {
+    convertedType: getType,
+    getOptionals
   },
   filesExtension: 'ts'
 };