+++ packages/templates/typescript/src/helpers/get-optionals.ts
@@ -1,0 +1,18 @@
+export function getOptionals(type, options) {
+  const config = options.data.root.config || {};
+
+  if (
+    config.avoidOptionals === '1' ||
+    config.avoidOptionals === 'true' ||
+    config.avoidOptionals === true ||
+    config.avoidOptionals === 1
+  ) {
+    return '';
+  }
+
+  if (!type.isRequired) {
+    return '?';
+  }
+
+  return '';
+}